tiled_catalog_broker/pyproject.toml
[project]
name = "vdp-broker"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "pyarrow",
    "h5py",
    "numpy",
    "ruamel.yaml",
    "canonicaljson",
    "sqlalchemy",
    "tiled[server]",
]

[project.scripts]
broker-inspect = "broker.cli:inspect_main"
broker-generate-yaml = "broker.cli:generate_yaml_main"
broker-ingest = "broker.cli:ingest_main"
broker-generate = "broker.cli:generate_main"
broker-register = "broker.cli:register_main"

[tool.hatch.build.targets.wheel]
packages = ["broker"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
